gdbstub: replace process ID 0 to a valid value
VELAPLATFO-52290

GDB uses id 0 to indicate any thread in RSP protocol. Previously use NuttX pid + 1 to avoid using ID 0.
Now we replace 0 with a non-zero value. In this way, the thread ID shows in GDB will match directly with
NuttX, makes it better for debugging. As for 0 which is the IDLE thread, we can easily idendify it too.

Change-Id: I5eab96e52d043362136837af8fd02ef2e372a33e
Signed-off-by: xuxingliang <xuxingliang@xiaomi.com>