arch: add register name/size to reginfo
VELAPLATFO-53264

These details are crucial for building a fully functional GDB stub.
Additionally, this enhancement simplifies the implementation of
gdbserver.py located in the tools/ directory. The inclusion of register
names specifically aids in matching the registers requested by GDB with
those available in the TCB, streamlining the debugging process.

Change-Id: I2498cf13056851902204aa9a2fc9242610ef1503
Signed-off-by: xuxingliang <xuxingliang@xiaomi.com>