nxgdb: fix info nxthreads to read current PC
VELAPLATFO-53048

Although GDB has a unifed register SP/PC for all architecture,
we need the real register name so we can read register from TCB's saved context.
GDB's python API can find all the registers but unfortunetly cannot be
used to distinguish which register is the real PC/SP, so we we have to
hardcode the name as a table using arch as key.

Change-Id: I9004960cbc1e4cebc750f837f2aef3e597b754f2
Signed-off-by: xuxingliang <xuxingliang@xiaomi.com>